# Faster init