--- a/test/cache-control.test.js
+++ b/test/cache-control.test.js
@@ -69,7 +69,7 @@ test('reply.preventCache API', t => {
     t.equal(res.statusCode, 200)
     t.equal(res.headers['cache-control'], 'no-store, max-age=0, private')
     t.equal(res.headers.pragma, 'no-cache')
-    t.equal(res.headers.expires, 0)
+    t.equal(res.headers.expires, '0')
     t.equal(res.payload, 'ok')
   })
 })