@Uzlopak 2023-07-27
Yes. Always avoid forEach map and co.